requirements.txt
Django==3.2.12
WeasyPrint==51
gunicorn==20.0.4
djangorestframework==3.12.4
django-cors-headers
dj-database-url==0.5.0
psycopg2-binary
